FIX: no longer cache the recommendations
Caching the results allows it to work with wire statements
however, the wire statements can no longer be used if recordId can be null.
meaning that this is now an option / required.

The issue this resolves is to ensure the badge is updated when the record is changed - without the user needing to leave the page. (Improving accuracy at the cost of performance)

Leaving this to true keeps this performant, but the results will no longer always be accurate.